add window for dialog and inventory

+++ core/src/main/java/io/src/model/MapModule/GameLocations/GameLocation.java
@@ -16,8 +16,8 @@ public class GameLocation extends Network {
 //    ArrayList<NPC> NPcs = new ArrayList<>();////
 
 
-    public Tile getTileByPosition(int x, int y){
-        return tiles[y][x];
+    public Tile getTileByPosition(float x, float y){
+        return tiles[(int)y][(int)x];
     }
 
     public Tile getTileByPosition(Position pos){
